코드 정리_1
