--- src/main/java/com/kh/servlet/FindUserpwServlet.java
@@ -1,34 +1,0 @@
-package com.kh.servlet;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-import com.kh.model.dao.UserDao;
-import com.kh.model.vo.User;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@WebServlet("/user/findpw")
-public class FindUserpwServlet extends HttpServlet {
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-
-            PrintWriter out = resp.getWriter();
-            String userId = req.getParameter("userId");
-
-            User user = new UserDao().findByUserId(userId);
-            
-            if (user != null) {
-                resp.setStatus(HttpServletResponse.SC_OK);
-                out.println(user.getUserPw());
-            } else {
-                resp.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-            }
-        }
-    
-}
