+++ src/main/java/com/kh/servlet/plan/DeletePlanServlet.java
@@ -1,41 +1,39 @@
-package com.kh.servlet;
-
-import java.io.IOException;
-import java.io.PrintWriter;
+package com.kh.servlet.plan;
 
 import com.kh.model.dao.PlanDao;
-
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
 
 /**
  * Servlet implementation class DeletePlanController
  */
 @WebServlet("/delete.pl")
 public class DeletePlanServlet extends HttpServlet {
-	
 
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		
-		request.setCharacterEncoding("utf-8");
-		response.setContentType("text/html;charset=utf-8");
-		response.setCharacterEncoding("utf-8");
-		String userId = "validUserId0";
-		String planIdS = request.getParameter("planId");
-		int planId = Integer.parseInt(planIdS);
-		int result = new PlanDao().deleteByPlanIdAndWriter(planId, userId);
-		
-		PrintWriter out = response.getWriter();
-		if (result != 0) {
+  protected void doGet(HttpServletRequest request, HttpServletResponse response)
+      throws ServletException, IOException {
+
+    request.setCharacterEncoding("utf-8");
+    response.setContentType("text/html;charset=utf-8");
+    response.setCharacterEncoding("utf-8");
+    String userId = "validUserId0";
+    String planIdS = request.getParameter("planId");
+    int planId = Integer.parseInt(planIdS);
+    int result = new PlanDao().deleteByPlanIdAndWriter(planId, userId);
+
+    PrintWriter out = response.getWriter();
+    if (result != 0) {
 //			request.getRequestDispatcher("main.html").forward(request, response);
-			response.setStatus(200);
-		} else {
-			response.setStatus(400);
-		}
-		out.close();
+      response.setStatus(200);
+    } else {
+      response.setStatus(400);
+    }
+    out.close();
 //		if(result != 0) {
 ////		List<Plan> list = new PlanDao().findByWriter(userId);
 //		HttpSession session = request.getSession();
@@ -44,9 +42,8 @@ public class DeletePlanServlet extends HttpServlet {
 //		} else {
 //			response.getWriter().print("삭제 실패했어요 ㅠㅠ");
 //		}
-		
-	}
 
+  }
 
 
 }
