+++ src/main/java/com/kh/servlet/user/SignupServlet.java
@@ -1,4 +1,4 @@
-package com.kh.servlet;
+package com.kh.servlet.user;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
