+++ src/main/resources/css/main.css
@@ -1,0 +1,184 @@
+body {
+    margin: 50px;
+    background-color: #f0f3f6;
+}
+
+/* Header */
+.mainTitle {
+    text-align: center;
+    font-weight: 900;
+}
+
+#headerIcons {
+    text-align: right;
+    align-items: right;
+    gap: 10px;
+}
+
+header .btn {
+    background-color: #f0f3f6;
+    border: 0px solid black;
+}
+
+/* main */
+
+main {
+    max-width: 800px;
+    margin: 20px auto;
+    padding: 20px;
+    background-color: white;
+    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+    border-radius: 10px;
+}
+
+main input,
+textarea {
+    margin-top: 5px;
+    width: 100%;
+    padding: 8px;
+    box-sizing: border-box;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    resize: vertical;
+}
+
+main button {
+    background-color: #f0f3f6;
+    color: black;
+    border: none;
+    cursor: pointer;
+    font-size: 30px;
+}
+
+main button:hover {
+    color: gray;
+}
+
+main a {
+    text-decoration: none;
+    color: black;
+}
+
+main ul {
+    list-style: none;
+    padding: 0;
+}
+
+main li {
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    margin-bottom: 10px;
+    cursor: pointer;
+}
+
+main input[type="date"] {
+    width: 35%;
+}
+
+main .btn {
+    background-color: #f0f3f6;
+    border: 0px solid black;
+}
+
+/* main .forming */
+
+.forming {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+#save {
+    width: 80px;
+    background-color: black;
+    color: #fff;
+    border-radius: 4px;
+}
+
+main strong {
+    display: flex;
+}
+
+/* main .completeBox */
+.completeBox {
+    display: flex;
+    border: 1px solid lightgray;
+    width: 100%;
+    display: flex;
+    border-radius: 4px;
+    text-align: center;
+    justify-content: space-evenly;
+    padding-top: 10px;
+    padding-bottom: 10px;
+}
+
+.completeBox span {
+    color: black;
+    font-size: 30px;
+}
+
+#title {
+    padding-left: 20px;
+    font-size: 15px;
+    border: 3px solid black;
+}
+
+/* main .time-box */
+.time-box {
+    display: flex;
+    justify-content: flex-start;
+}
+
+.endAlarm {
+    align-items: flex-end;
+}
+
+#endAlarmDateBoolean {
+    width: 15px;
+    margin-right: 5px;
+}
+
+/* main #planSection */
+#planSection {
+    margin-top: 20px;
+    align-items: center;
+    text-align: center;
+}
+
+#planSectionHeader {
+    display: flex;
+    justify-content: space-between;
+    color: black;
+    font-size: larger;
+}
+
+/* main -> main.js */
+.plannerItem {
+    padding-left: 20px;
+    font-size: 15px;
+    border: 1px solid lightgray;
+    background-color: #fff;
+    height: 40.5px;
+    display: flex;
+    justify-content: space-between;
+    padding: 10px;
+    background-color: #f9f9f9;
+    border-bottom: 1px solid #ccc;
+}
+
+.plannerDate {
+    color: #777;
+    text-align: end;
+    flex: 1;
+}
+
+.deleteButton {
+    color: #e91e63;
+    cursor: pointer;
+}
+
+.comRadio {
+    width: 30px;
+    margin-right: 10px;
+    margin-left: -10px;
+}
