--- src/main/resources/css/plan.css
@@ -1,268 +1,0 @@
-body {
-
-    margin: 50px;
-    background-color: #f0f3f6;
-}
-
-
-h1 {
-
-    text-align: center;
-    font-weight: 900;
-}
-
-#icons {
-    text-align: right;
-    align-items: right;
-    gap: 10px;
-
-}
-
-#headerIcons button {
-    padding: 5px;
-    cursor: pointer;
-}
-
-main {
-    max-width: 800px;
-    margin: 20px auto;
-    padding: 20px;
-    background-color: white;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-    border-radius: 10px;
-}
-
-.forming {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-}
-
-label {
-    margin-top: 10px;
-
-}
-
-input,
-textarea {
-    margin-top: 5px;
-    width: 100%;
-    padding: 8px;
-    box-sizing: border-box;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    resize: vertical;
-}
-
-button {
-    background-color: #f0f3f6;
-    color: black;
-    border: none;
-    cursor: pointer;
-    font-size: 30px;
-}
-
-button:hover {
-
-    color: gray;
-}
-
-#plannerList {
-    margin-top: 20px;
-    align-items: center;
-    text-align: center;
-}
-
-ul {
-    list-style: none;
-    padding: 0;
-}
-
-li {
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    margin-bottom: 10px;
-    cursor: pointer;
-}
-
-.plannerItem {
-    display: flex;
-    justify-content: space-between;
-    padding: 10px;
-    background-color: #f9f9f9;
-    border-bottom: 1px solid #ccc;
-}
-
-.plannerDate {
-    color: #777;
-    
-    text-align: end;
-    flex: 1;
-}
-
-.deleteButton {
-    color: #e91e63;
-    cursor: pointer;
-}
-
-.plannerDetails {
-    display: none;
-    padding: 10px;
-    background-color: #f0f0f0;
-}
-
-.showDetails {
-    display: block;
-}
-
-.dark-mode {
-    background-color: #1a1a1a;
-    color: #ffffff;
-}
-
-.toggleSwitch {
-    width: 100px;
-    height: 50px;
-    display: block;
-    position: relative;
-    border-radius: 30px;
-    background-color: #fff;
-    box-shadow: 0 0 16px 3px rgba(0 0 0 / 15%);
-    cursor: pointer;
-    margin: 30px;
-}
-
-.plan-complete {
-	text-decoration: line-through;
-	text-decoration-thickness: 3px;
-}
-
-a {
-	text-decoration: none;
-    color: black;
-}
-
-
-.complete_box {
-    display: flex;
-    border: 1px solid lightgray;
-    width: 100%;
-    display: flex;
-    border-radius: 4px;
-    text-align: center;
-    justify-content: space-evenly;
-    padding-top: 10px;
-    padding-bottom: 10px;
-
-}
-
-.not-completed,
-.completed {
-    color: black;
-    font-size: 30px;
-}
-
-#title {
-    padding-left: 20px;
-    font-size: 15px;
-    border : 3px solid black;
-}
-
-
-.time-box {
-    display: flex;
-    justify-content: flex-start;
-
-}
-
-input[type="date"] {
-    width: 35%;
-}
-
-.endAlarm {
-    align-items: flex-end;
-}
-
-#endAlarmDateBoolean {
-    width: 15px;
-    margin-right: 5px;
-}
-
-#save {
-    width: 80px;
-    background-color: black;
-    color: #fff;
-    border-radius: 4px;
-
-}
-
-.plannerItem {
-    padding-left: 20px;
-    font-size: 15px;
-    border: 1px solid lightgray;
-    background-color: #fff;
-    height: 40.5px;
-}
-
-strong {
-    display: flex;
-}
-
-
-
-.plannerDetails {
-    background-color: #f0f3f6;
-}
-
-#toDetailDiv {
-    display: flex;
-    justify-content: end;
-}
-
-#toDetailButton {
-    text-align: center;
-    bottom: 5px;
-    right: 5px;
-    padding: 5px;
-    cursor: pointer;
-    border: 1px solid white;
-    color: white;
-    border-radius: 4px;
-    width: 80px;
-    font-size: 15px;
-    text-align: center;
-    background-color: black;
-}
-
-#planLists {
-    display: flex;
-    justify-content: space-between;
-    color: black;
-    font-size: larger;
-}
-
-.btn3 {
-    background-color: #fff;
-    margin-bottom: 15px;
-    height: 30px;
-}
-
-.com_radio {
-    width: 30px;
-    margin-right: 10px;
-    margin-left: -10px;
-}
-
-.dropdown-menu-profile {
-    position: absolute;
-    right: 10px;
-    height: 170px;
-    border-radius: 1px;
-}
-
-.btn {
-    background-color: #f0f3f6;
-    border: 0px solid black;
-
-}
