+++ src/main/resources/img/planner_main.png
