+++ src/main/resources/index.html
@@ -2,14 +2,14 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <link href="planner_favicon.png" rel="icon" type="image/x-icon"/>
+  <link href="img/planner_favicon.png" rel="icon" type="image/x-icon"/>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Planer</title>
   <link rel="stylesheet" href="css/style.css">
 </head>
 <body>
 <h1>Planer</h1><br>
-<img width="250px" height="250px" src="plr.png" alt="플래너 이미지">
+<img width="250px" height="250px" src="img/planner_main.png" alt="플래너 이미지">
 <button onclick="location.href='user/signin.html'">로그인</button>
 <button onclick="location.href='user/signup.html'">회원가입</button>
 </body>
